rename: 과제 관련 타입 파일명 수정 #11
